chore: ChallengeTitle에서 ChallengeReward로 변경
